- ISWCへ持っていくための準備その１． - eclipse3 M4でコンパイラの設定を厳しめに設定して，不要なキャストなどを取り除いた．
git-svn-id: svn+ssh://svn.code.sf.net/p/mr3/code@211 2d18f6ef-ed67-4be4-865d-64751944eb35

+++ mr3_alpha8_to_9/trunk/MR3_from8/src/jp/ac/shizuoka/cs/panda/mmm/mr3/actions/ZoomAction.java
@@ -68,9 +68,9 @@ public class ZoomAction extends AbstractAction {
 				return;
 			}
 			if (Math.abs(s.getWidth() - (p.x + p.getWidth())) > Math.abs(s.getHeight() - (p.x + p.getHeight())))
-				scale = (double) s.getWidth() / (p.x + p.getWidth());
+				scale = s.getWidth() / (p.x + p.getWidth());
 			else
-				scale = (double) s.getHeight() / (p.y + p.getHeight());
+				scale = s.getHeight() / (p.y + p.getHeight());
 			scale = Math.max(Math.min(scale, 16), .01);
 			graph.setScale(scale);
 		}
